downstream/requirements.txt
einops==0.7.0
h5py==3.8.0
MedPy==0.4.0
numpy==1.21.6
Pillow==10.3.0
scipy==1.7.3
torch==1.8.1+cu111
torchvision==0.9.1+cu111
tqdm==4.66.2
